+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixed
 
+- Csv: fix get-metadatas from CSVs files with `skiprows` as list (0-indexed) in Datasource.
 - FTP: retry connection on `SSHException` while opening a remote url.
 
 ## [0.12.0] - 2023-09-01